Renamed controller/auth.js by  controller/user.js
